Refactored solution to be contained within ActiveQuery instead of helper
